+++ youtui/src/app/ui/action.rs
@@ -1,3 +1,5 @@
+use crate::app::component::actionhandler::Action;
+
 #[derive(Clone, PartialEq, Default, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum AppAction {
